pyproject.toml
[tool.poetry]
name = "task-manager"
version = "0.1.0"
description = ""
authors = ["DorLex <dorozhinskiysanya@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django-ninja = "^1.1.0"
python-dotenv = "^1.0.1"
uvicorn = "^0.29.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
